unit tests, some refactoring :
