+++ BlazorMarkDownAppJwt/BlazorMarkDownAppJwt/Server/Services/MarkDowns/IDocumentService.cs
@@ -5,6 +5,9 @@ namespace BlazorMarkDownAppJwt.Server.Services.MarkDowns
     public interface IDocumentService
     {
         Task<Document?> GetDocument();
+
         Task<Document?> UpsertDocument(string markDown);
+
+        Task<Document?> GetReadMeDocument(string markDown);
     }
 }
