+++ BlazorMarkDownAppJwt/BlazorMarkDownAppJwt/Shared/Docs/DevNotes.txt
@@ -1,4 +1,7 @@
-﻿JWT Token authent documentation - in Blazor
+﻿asp.net core 6
+https://docs.microsoft.com/fr-fr/aspnet/core/?view=aspnetcore-6.0
+
+JWT Token authent documentation - in Blazor
 https://www.youtube.com/watch?v=bu6F4oeYdzA
 https://docs.microsoft.com/fr-fr/aspnet/core/blazor/security/webassembly/hosted-with-identity-server?view=aspnetcore-6.0&tabs=visual-studio
 https://www.prowaretech.com/articles/current/blazor/wasm/jwt-authentication-simple
@@ -23,3 +26,6 @@ https://docs.microsoft.com/fr-fr/aspnet/core/web-api/action-return-types?view=as
 
 unit tests : 
 https://stackoverflow.com/questions/54219742/mocking-ef-core-dbcontext-and-dbset
+
+examples of Apps
+https://stackedit.io/
